requirements.txt
google-cloud-datastore==1.7.3
simplejson